+++ test/runtests.jl
@@ -10,6 +10,7 @@ Random.seed!(0x5c1e9a3b)
 
 include("exact_tests.jl")
 include("spaces_tests.jl")
+include("tensorspaces_tests.jl")
 include("brackets_tests.jl")
 include("algebras_tests.jl")
 include("torus_tests.jl")